pyproject.toml
[project]
name = "pixellm"
version = "0.1.0"
description = "Teaching small LLMs to draw pixel art via code generation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.4",
    "numpy>=2.4.4",
    "peft>=0.19.1",
    "pillow>=12.2.0",
    "pydantic>=2.13.2",
    "transformers>=5.5.4",
    "trl>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pixellm"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
